+++ lib/src/home/pages/answer/answer_bloc.dart
@@ -1,0 +1,9 @@
+import 'package:bloc_pattern/bloc_pattern.dart';
+
+class AnswerBloc extends BlocBase {
+  //dispose will be called automatically by closing its streams
+  @override
+  void dispose() {
+    super.dispose();
+  }
+}
